Disable the nightly formatter by default
Hi! I saw the default nightly formatting feature merge earlier and I realised that I was kind of hosed. Why? I don't use Rustup, and this feature relies on Rustup, so my kraken fmt doesn't work anymore.

I'm also a bit concerned that the behaviour as implemented forces every dev team onto nightly Rust, which I don't think we should be encouraging without careful thought and some tooling.

I'd therefore like us to take a quick step back by making it be an opt-in and not an opt-out feature so as to give more time for the above attributes to be addressed. Sound good?