Make Rust builds respect Cargo.lock when present (#202)
* feature: Make Rust builds respect Cargo.lock when present

Kraken should, by default, respect the `Cargo.lock` present in the
current Cargo project rather than silently pull down dependencies not
explicitly noted in the initial `Cargo.lock`. This is accomplished by
passing `--locked` to `cargo. If the lockfile is out of date, the
underlying package should be updated to reflect that change if
`Cargo.lock` is checked in.

To avoid breaking existing workflows that don't check in `Cargo.lock`,
the default behavior is to only pass `--locked` if a `Cargo.lock` exists
in the parent directory tree of the current working directory.

The flag only really matters to CargoBuildTask, CargoTestTask, and
CargoMetadata, but is also propagated to CargoPublishTask so that it
isn't allowed to silently update the lockfile if it happens to be called
first. Clippy doesn't take `--locked`, and so remains unmodified.

Note also that checking in `Cargo.lock` is now recommended by the Rust
project: https://blog.rust-lang.org/2023/08/29/committing-lockfiles.html

* Update kraken-build/src/kraken/std/cargo/manifest.py

---------

Co-authored-by: Niklas Rosenstein <rosensteinniklas@gmail.com>